requirements.txt
streamlit
pandas
altair
numpy
matplotlib.pyplot